move pkg/* to internal/*
these are the packages that aren't meant for external applications.
config/* must die in favour of /env/
