+++ internal/sftpclient/sftpclient.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/fusor/cpma/pkg/config"
+	"github.com/fusor/cpma/internal/config"
 	"github.com/pkg/sftp"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/crypto/ssh"
